修改collect task的结构

+++ agent-manager-v2/agent-manager-common/src/main/java/com/didichuxing/datachannel/agentmanager/common/bean/domain/agent/AgentMetricDO.java
@@ -25,7 +25,7 @@ public class AgentMetricDO extends BaseDO {
 
     private Integer fdCount;
 
-    private Integer limitTps;
+    private Long limitTps;
 
     private Date startTime;
 
@@ -113,11 +113,11 @@ public class AgentMetricDO extends BaseDO {
         this.fdCount = fdCount;
     }
 
-    public Integer getLimitTps() {
+    public Long getLimitTps() {
         return limitTps;
     }
 
-    public void setLimitTps(Integer limitTps) {
+    public void setLimitTps(Long limitTps) {
         this.limitTps = limitTps;
     }
